Split backend sync workflows
